+++ feature/dailyduas/src/main/java/com/raydev/dailyduas/api/DailyDuasModel.kt
@@ -1,4 +1,4 @@
-package com.raydev.dailyduas
+package com.raydev.dailyduas.api
 
 /**
  * @author Raihan Arman
