+++ feature/dailyduas/src/main/java/com/raydev/dailyduas/api/FirestoreClient.kt
@@ -1,0 +1,16 @@
+package com.raydev.dailyduas.api
+
+import kotlinx.coroutines.flow.Flow
+
+/**
+ * @author Raihan Arman
+ * @date 20/05/24
+ */
+sealed class FirestoreClientResult {
+    data class Success(val root: List<DailyDuasModel>) : FirestoreClientResult()
+    data class Failure(val exception: Exception) : FirestoreClientResult()
+}
+
+interface FirestoreClient {
+    fun getDailyDuas(): Flow<FirestoreClientResult>
+}
