+++ feature/dailyduas/src/main/java/com/raydev/dailyduas/domain/DomainException.kt
@@ -1,4 +1,4 @@
-package com.raydev.dailyduas
+package com.raydev.dailyduas.domain
 
 /**
  * @author Raihan Arman
